requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.100.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.17
pulp==2.7.0
streamlit==1.24.0
schedule==1.2.0
python-dotenv==1.0.0
pytest==7.4.0

